+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
-### Fixed
-- Fixed a test that broke when switching to latest EAP version (groovy resolving issue)
+## [1.5.8]
+### Changed
+- [#3] Returns a multi reference for services in case of multiple implementations for a single service name.
 
 ## [1.5.7]
 ### Added